// types/ethers-contracts/ProofLayerCertificateRegistry.ts
export declare namespace ProofLayerCertificateRegistry {
      
    export type CertificateStruct = {certificateId: BytesLike, assetId: BytesLike, claimType: BytesLike, policyId: BytesLike, evidenceRoot: BytesLike, observedAt: BigNumberish, validUntil: BigNumberish, independentRootCount: BigNumberish, result: BigNumberish, issuer: AddressLike, revoked: boolean}

    export type CertificateStructOutput = [certificateId: string, assetId: string, claimType: string, policyId: string, evidenceRoot: string, observedAt: bigint, validUntil: bigint, independentRootCount: bigint, result: bigint, issuer: string, revoked: boolean] & {certificateId: string, assetId: string, claimType: string, policyId: string, evidenceRoot: string, observedAt: bigint, validUntil: bigint, independentRootCount: bigint, result: bigint, issuer: string, revoked: boolean }
  
    }